[cxx-interop] Fix wrong field offsets with value types using inheritance
The code path responsible for calculating field offsets creates opaque
fields for all the base classes. This code path simply summed up the
sizes of the base classes to get the offset. Unfortunately, this does
not work for base classes with vptrs. C++ does not embedd the vptrs from
the base classes into the derived class. The derived class has a single
vptr. This PR removes this assumption that we can just add the base
class sizes up and relies purely on base and field offsets to calculate
the sizes for the opaque fields representing the base classes.